// frontend/lib/api/generated/model/pageGuiaRemisionListOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { GuiaRemisionListOut } from './guiaRemisionListOut';

export interface PageGuiaRemisionListOut {
  items: GuiaRemisionListOut[];
  total: number;
  page: number;
  page_size: number;
}